refactor(#12): 공용 파서 테스트 분리
