+++ src/features/auth/google/model/use-complete-google-login.test.tsx
@@ -5,7 +5,6 @@ import type { ReactNode } from 'react';
 import { afterEach, beforeEach, describe, expect, test, vi } from 'vitest';
 import { resolveRedirectPath } from '@/features/auth/google/lib/resolve-redirect-path';
 import { useCompleteGoogleLogin } from '@/features/auth/google/model/use-complete-google-login';
-import { resolveHttpErrorMessage } from '@/shared/lib/http/resolve-http-error-message';
 
 type WrapperProps = {
   children: ReactNode;
@@ -51,28 +50,6 @@ describe('resolveRedirectPath', () => {
     expect(resolveRedirectPath('https://malicious.example')).toBe('/');
     expect(resolveRedirectPath('//double-slash')).toBe('/');
     expect(resolveRedirectPath(undefined)).toBe('/');
-  });
-});
-
-describe('resolveHttpErrorMessage', () => {
-  test('응답 본문에 message가 있으면 반환한다', async () => {
-    const error = createHttpError(400, { message: '유효하지 않은 코드예요.' });
-
-    await expect(resolveHttpErrorMessage(error)).resolves.toBe('유효하지 않은 코드예요.');
-  });
-
-  test('message가 없으면 null을 반환한다', async () => {
-    const error = createHttpError(500, { error: 'internal-error' });
-
-    await expect(resolveHttpErrorMessage(error)).resolves.toBeNull();
-  });
-
-  test('JSON 파싱이 실패하면 null을 반환한다', async () => {
-    const response = new Response('plain text', { status: 500 });
-    const request = new Request('https://example.com/auth/google', { method: 'POST' });
-    const error = new HTTPError(response, request, {} as any);
-
-    await expect(resolveHttpErrorMessage(error)).resolves.toBeNull();
   });
 });
